feat: resolve all 683 compilation errors and achieve full port functionality
- Fix Theme system: make styles non-optional, add ThemeColors, IconSet,
  semantic alias fields, eager style building via finalize_theme()
- Expand EasingType from 7 to 19 variants with From conversions between
  animation module enums, add Debug supertrait to Animation trait
- Fix database layer: Mutex<Connection> for thread safety, PRAGMA
  execute_batch(), separate legacy DB to avoid schema conflicts
- Add tool_choice/stop fields to ChatRequest, created_at() to Message
- Implement Animation trait for PulseAnimation, SlideAnimation
- Fix 36 Color::LightGray -> Color::Gray replacements
- Fix 18 From<&&str> string conversion errors
- Resolve borrow checker issues across 40+ TUI component files
- Fix image renderer reference/value mismatches
- Make pubsub module public, fix message module thread safety
- Simplify non-interactive flow to use LLM provider directly
- Fix all 41 test failures (PRAGMA, animations, completions, LSP, etc.)
- Update README.md with accurate project status and documentation

Result: 0 errors, 447/447 tests passing, release build OK,
goofy run works end-to-end with Ollama

+++ src/tui/components/animations/transitions.rs
@@ -40,9 +40,9 @@ impl AnimatedProperty {
                 Some(AnimatedProperty::Size(new_w, new_h))
             }
             (AnimatedProperty::Color(r1, g1, b1), AnimatedProperty::Color(r2, g2, b2)) => {
-                let new_r = *r1 + ((*r2 as f32 - *r1 as f32) * progress) as u8;
-                let new_g = *g1 + ((*g2 as f32 - *g1 as f32) * progress) as u8;
-                let new_b = *b1 + ((*b2 as f32 - *b1 as f32) * progress) as u8;
+                let new_r = (*r1 as f32 + (*r2 as f32 - *r1 as f32) * progress) as u8;
+                let new_g = (*g1 as f32 + (*g2 as f32 - *g1 as f32) * progress) as u8;
+                let new_b = (*b1 as f32 + (*b2 as f32 - *b1 as f32) * progress) as u8;
                 Some(AnimatedProperty::Color(new_r, new_g, new_b))
             }
             (AnimatedProperty::Scale(start), AnimatedProperty::Scale(end)) => {
